+++ components/esp_peripherals/periph_wifi.c
@@ -327,7 +327,7 @@ static void _wifi_event_callback(void *arg, esp_event_base_t event_base,
         esp_periph_start_timer(self, periph_wifi->reconnect_timeout_ms / portTICK_RATE_MS, wifi_reconnect_timer);
 
     } else {
-        ESP_LOGW(TAG, "WiFi Event cb, Unhandle event_base:%s, event_id:%d", event_base, event_id);
+        ESP_LOGW(TAG, "WiFi Event cb, Unhandle event_base:%s, event_id:%d", event_base, (int)event_id);
     }
 }
 #else
